rearrange packages and move websocket message broker auth to a separate class
