+++ websocket/src/main/java/com/success/websocket/security/WebsocketMessageBrokerAuthentication.java
@@ -1,4 +1,4 @@
-package com.success.websocket;
+package com.success.websocket.security;
 
 import java.util.Collections;
 import java.util.List;
@@ -9,7 +9,6 @@ import org.springframework.core.annotation.Order;
 import org.springframework.messaging.Message;
 import org.springframework.messaging.MessageChannel;
 import org.springframework.messaging.simp.config.ChannelRegistration;
-import org.springframework.messaging.simp.config.MessageBrokerRegistry;
 import org.springframework.messaging.simp.stomp.StompCommand;
 import org.springframework.messaging.simp.stomp.StompHeaderAccessor;
 import org.springframework.messaging.support.ChannelInterceptor;
@@ -17,7 +16,6 @@ import org.springframework.messaging.support.MessageHeaderAccessor;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.Authentication;
 import org.springframework.web.socket.config.annotation.EnableWebSocketMessageBroker;
-import org.springframework.web.socket.config.annotation.StompEndpointRegistry;
 import org.springframework.web.socket.config.annotation.WebSocketMessageBrokerConfigurer;
 
 import lombok.extern.slf4j.Slf4j;
@@ -25,26 +23,10 @@ import lombok.extern.slf4j.Slf4j;
 @Slf4j
 @Configuration
 @EnableWebSocketMessageBroker
-//we want web socket security takes precedence over web security to allow the inbound connections after validating the tokens...
-@Order(Ordered.HIGHEST_PRECEDENCE) 
-public class WebsocketMessageBroker implements WebSocketMessageBrokerConfigurer {
-
-  @Override
-  public void configureMessageBroker(MessageBrokerRegistry registry) {
-    // handles the outgoing messages intended for destination "/topic", /user, etc.
-    // clients subscribe to this destination to receive message
-    registry.enableSimpleBroker("/topic", "/user");
-    // client use this prefix.. for e.g. /myprefix/incoming.. where "incoming" is mapped to
-    // MessageHandler
-    registry.setApplicationDestinationPrefixes("/myprefix", "/chat");
-    // for user to user messages
-    registry.setUserDestinationPrefix("/user");
-  }
-
-  @Override
-  public void registerStompEndpoints(StompEndpointRegistry registry) {
-    registry.addEndpoint("/myendpoint").withSockJS();
-  }
+// we want web socket security takes precedence over web security to allow the inbound connections
+// after validating the tokens...
+@Order(Ordered.HIGHEST_PRECEDENCE)
+public class WebsocketMessageBrokerAuthentication implements WebSocketMessageBrokerConfigurer {
 
   public void configureClientInboundChannel(ChannelRegistration registration) {
     registration.interceptors(
